+++ dotnet/src/AuthTokenHandling.WinFormsClient/OAuthTokenRequestor.cs
@@ -1,6 +1,6 @@
 using CefSharp.DevTools.Database;
 using CefSharp.DevTools.DOM;
-using Newtonsoft.Json.Linq;
+//using Newtonsoft.Json.Linq;
 using System;
 using System.Collections.Generic;
 using System.Linq;